Bitmap Converter for the lcdgfx library - multiple bitmaps edition
The lcdgfx library (https://github.com/lexus2k/lcdgfx) can handle up to 16 Bit RGB565 bitmaps. 
However, the library expects a little endian encoded uint_8 array of the bitmap while most of
the bitmap convertes (for example http://javl.github.io/image2cpp/) are providing a uint_16 
encoded output.

Due to the limited ressources of the desired hardware, it seems to be difficult and not very 
usefull to convert the data during runtime. This PS script does the job on ever Windows based 
PC. For other plattforms please feel free to adopt the process. 

THIS VERSIONS is for the conversation of multiple bitmaps in one file / output! 
For example numbers or glyphs.